+++ package.json
@@ -12,7 +12,6 @@
 		"react-native": "^0.49.3",
 		"react-native-browser": "^0.6.2",
 		"react-native-elements": "^0.18.2",
-		"react-native-optimized-flatlist": "^1.0.3",
 		"react-native-router-flux": "^4.0.0-beta.22",
 		"react-native-tab-view": "^0.0.70",
 		"react-native-vector-icons": "^4.4.2",
